+++ src/main/java/roomescape/controller/api/ReservationApiController.java
@@ -4,8 +4,8 @@ import jakarta.validation.Valid;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-import roomescape.controller.dto.ReservationRequest;
-import roomescape.controller.dto.ReservationResponse;
+import roomescape.controller.dto.reservation.ReservationRequest;
+import roomescape.controller.dto.reservation.ReservationResponse;
 import roomescape.entity.Reservation;
 import roomescape.service.ReservationService;
