--- src/main/java/roomescape/controller/dto/AvailableReservationTimeSearchRequest.java
@@ -1,9 +1,0 @@
-package roomescape.controller.dto;
-
-import java.time.LocalDate;
-
-public record AvailableReservationTimeSearchRequest(
-        LocalDate date,
-        long themeId
-) {
-}
