+++ src/test/java/roomescape/controller/api/ReservationTimeApiTest.java
@@ -6,8 +6,8 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.annotation.DirtiesContext;
-import roomescape.controller.dto.ReservationTimeRequest;
-import roomescape.controller.dto.ThemeRequest;
+import roomescape.controller.dto.reservationTime.ReservationTimeRequest;
+import roomescape.controller.dto.theme.ThemeRequest;
 
 import java.time.LocalDate;
 import java.util.HashMap;
